Code cleanup (#17237)
* renames in dbengine

* remove leftovers from memory mode save and map

* fix docs about 3 tiers by default

* split linked-lists, bitmaps and storage-points from libnetdata.h

(cherry picked from commit 00f897a883bde68ba3df587655a30225fbc4ecdd)
