+++ src/database/rrdhost.c
@@ -360,12 +360,6 @@ int is_legacy = 1;
 
     rrdset_index_init(host);
 
-    if(config_get_boolean(CONFIG_SECTION_DB, "delete obsolete charts files", 1))
-        rrdhost_option_set(host, RRDHOST_OPTION_DELETE_OBSOLETE_CHARTS);
-
-    if(config_get_boolean(CONFIG_SECTION_DB, "delete orphan hosts files", 1) && !is_localhost)
-        rrdhost_option_set(host, RRDHOST_OPTION_DELETE_ORPHAN_HOST);
-
     char filename[FILENAME_MAX + 1];
     if(is_localhost)
         host->cache_dir  = strdupz(netdata_configured_cache_dir);
